Move common functions to utils
